Changed the order of chunk locking during generation

+++ VoxelEngine/src/Terrain/Chunk.cpp
@@ -162,7 +162,7 @@ void Chunk::AddEdgeMesh(VoxelMeshBuilder &meshBuilder, Voxel &v, VoxelFace f1, V
 
 void Chunk::DetermineVoxelFeatures(Voxel &v, size_t x, size_t z, size_t h)
 {
-    int y = &v - &m_VoxelGrid[x][z][0];
+    size_t y = &v - &m_VoxelGrid[x][z][0];
     if (y >= h)
         return;
     double density = m_Perlin.octave3D((m_Position.x * CHUNK_WIDTH + x) * 0.02,
